see [8](https://github.com/Islandora-Labs/islandora_vagrant_base_box/issues/8)
Matching drush version with what is currently in travis for release. Travis would need to be updated and ensure no tests fail before we can move base box to drush 7. This should be considered for next release but it is out of scope of islandora_vagrant_base_box issue. We should always match community deployment guide and vagrant box drush version to what is currently in Travis.

To test bring up vagrant box based on pull and ensure drush version is 6.3.0 and that you can run drush commands as expected.